Do not wait handling of incoming webhook (#14)
Handling webhooks may take longer that the caller's request timeout (suspected to be 100 seconds). Azure DevOps seeps to disabled (labeled as `Enabled (restricted)`) the service hook when one (may be two) request is timed out which can make one wonder why new webhooks are not getting handled.

This PR changes the endpoint handler so that the process is not awaited.